[CHORE] Move Font Settings to Separate Menu
