+++ src/features/island/hud/components/settings-menu/general-settings/AppearanceSettings.tsx
@@ -1,16 +1,14 @@
 import React from "react";
 import { Button } from "components/ui/Button";
-import { Label } from "components/ui/Label";
 import { useIsDarkMode } from "lib/utils/hooks/useIsDarkMode";
 import { useAppTranslation } from "lib/i18n/useAppTranslations";
-import { Font, changeFont } from "lib/utils/fonts";
+import { ContentComponentProps } from "../GameOptions";
 
-export const AppearanceSettings: React.FC = () => {
+export const AppearanceSettings: React.FC<ContentComponentProps> = ({
+  onSubMenuClick,
+}) => {
   const { t } = useAppTranslation();
   const { isDarkMode, toggleDarkMode } = useIsDarkMode();
-  const currentLanguage = localStorage.getItem("language") || "en";
-
-  const fonts: Font[] = ["Default", "Bold", "Sans Serif", "Chunky (Old)"];
 
   return (
     <>
@@ -21,19 +19,9 @@ export const AppearanceSettings: React.FC = () => {
             : t("gameOptions.generalSettings.lightMode")}
         </span>
       </Button>
-      <Label type="default" className="mb-2">
+      <Button className="mb-1" onClick={() => onSubMenuClick("font")}>
         {t("gameOptions.generalSettings.font")}
-      </Label>
-      {fonts.map((font) => (
-        <Button
-          key={font}
-          className="mb-1"
-          disabled={currentLanguage === "zh-CN"}
-          onClick={() => changeFont(font)}
-        >
-          <span>{font}</span>
-        </Button>
-      ))}
+      </Button>
     </>
   );
 };
